+++ backend/src/baserow/api/groups/serializers.py
@@ -2,9 +2,14 @@ from rest_framework import serializers
 
 from baserow.core.models import Group
 
-from .users.serializers import GroupUserGroupSerializer
+from .users.serializers import GroupUserGroupSerializer, GroupUserSerializer
 
-__all__ = ["GroupUserGroupSerializer", "GroupSerializer", "OrderGroupsSerializer"]
+__all__ = [
+    "GroupUserGroupSerializer",
+    "GroupSerializer",
+    "OrderGroupsSerializer",
+    "GroupWithUsersSerializer",
+]
 
 
 class GroupSerializer(serializers.ModelSerializer):
@@ -13,6 +18,21 @@ class GroupSerializer(serializers.ModelSerializer):
         fields = (
             "id",
             "name",
+        )
+        extra_kwargs = {"id": {"read_only": True}}
+
+
+class GroupWithUsersSerializer(serializers.ModelSerializer):
+    users = GroupUserSerializer(
+        many=True, source="groupuser_set", required=False, read_only=True
+    )
+
+    class Meta:
+        model = Group
+        fields = (
+            "id",
+            "name",
+            "users",
         )
         extra_kwargs = {"id": {"read_only": True}}