+++ backend/src/baserow/api/groups/views.py
@@ -38,7 +38,6 @@ from baserow.core.models import GroupUser
 from baserow.core.trash.exceptions import CannotDeleteAlreadyDeletedItem
 
 from .errors import ERROR_GROUP_USER_IS_LAST_ADMIN
-from .schemas import group_user_schema
 from .serializers import GroupSerializer, OrderGroupsSerializer
 
 
@@ -56,12 +55,16 @@ class GroupsView(APIView):
             "are custom for each user. The order is configurable via the "
             "**order_groups** endpoint."
         ),
-        responses={200: build_array_type(group_user_schema)},
+        responses={200: GroupUserGroupSerializer(many=True)},
     )
     def get(self, request):
         """Responds with a list of serialized groups where the user is part of."""
 
-        groups = GroupUser.objects.filter(user=request.user).select_related("group")
+        groups = (
+            GroupUser.objects.filter(user=request.user)
+            .select_related("group")
+            .prefetch_related("group__users")
+        )
         serializer = GroupUserGroupSerializer(groups, many=True)
         return Response(serializer.data)
 
@@ -75,7 +78,7 @@ class GroupsView(APIView):
             "created via other endpoints."
         ),
         request=GroupSerializer,
-        responses={200: group_user_schema},
+        responses={200: GroupUserGroupSerializer},
     )
     @transaction.atomic
     @validate_body(GroupSerializer)
@@ -160,7 +163,7 @@ class GroupView(APIView):
         ),
         request=GroupSerializer,
         responses={
-            200: group_user_schema,
+            204: None,
             400: get_error_schema(
                 [
                     "ERROR_USER_NOT_IN_GROUP",