+++ backend/src/baserow/ws/signals.py
@@ -2,10 +2,67 @@ from django.db import transaction
 from django.dispatch import receiver
 
 from baserow.api.applications.serializers import get_application_serializer
-from baserow.api.groups.serializers import GroupSerializer, GroupUserGroupSerializer
+from baserow.api.groups.serializers import (
+    GroupSerializer,
+    GroupUserGroupSerializer,
+    GroupUserSerializer,
+)
+from baserow.api.user.serializers import PublicUserSerializer
 from baserow.core import signals
+from baserow.core.models import GroupUser
 
-from .tasks import broadcast_to_group, broadcast_to_users
+from .tasks import broadcast_to_group, broadcast_to_groups, broadcast_to_users
+
+
+@receiver(signals.user_updated)
+def user_updated(sender, performed_by, user, **kwargs):
+    group_ids = [
+        group_user.group_id for group_user in GroupUser.objects.filter(user=user)
+    ]
+
+    transaction.on_commit(
+        lambda: broadcast_to_groups.delay(
+            group_ids,
+            {"type": "user_updated", "user": PublicUserSerializer(user).data},
+            getattr(performed_by, "web_socket_id", None),
+        )
+    )
+
+
+@receiver(signals.user_deleted)
+def user_deleted(sender, performed_by, user, **kwargs):
+    group_ids = [
+        group_user.group_id for group_user in GroupUser.objects.filter(user=user)
+    ]
+
+    transaction.on_commit(
+        lambda: broadcast_to_groups.delay(
+            group_ids, {"type": "user_deleted", "user": PublicUserSerializer(user).data}
+        )
+    )
+
+
+@receiver(signals.user_restored)
+def user_restored(sender, performed_by, user, **kwargs):
+    group_ids = [
+        group_user.group_id for group_user in GroupUser.objects.filter(user=user)
+    ]
+
+    transaction.on_commit(
+        lambda: broadcast_to_groups.delay(
+            group_ids,
+            {"type": "user_restored", "user": PublicUserSerializer(user).data},
+        )
+    )
+
+
+@receiver(signals.user_permanently_deleted)
+def user_permanently_deleted(sender, user_id, group_ids, **kwargs):
+    transaction.on_commit(
+        lambda: broadcast_to_groups.delay(
+            group_ids, {"type": "user_permanently_deleted", "user_id": user_id}
+        )
+    )
 
 
 @receiver(signals.group_created)
@@ -45,19 +102,63 @@ def group_deleted(sender, group_id, group, group_users, user=None, **kwargs):
     )
 
 
-@receiver(signals.group_user_updated)
-def group_user_updated(sender, group_user, user, **kwargs):
+@receiver(signals.group_user_added)
+def group_user_added(sender, group_user, user, **kwargs):
     transaction.on_commit(
-        lambda: broadcast_to_users.delay(
-            [group_user.user_id],
+        lambda: broadcast_to_group.delay(
+            group_user.group_id,
             {
-                "type": "group_updated",
+                "type": "group_user_added",
+                "id": group_user.id,
                 "group_id": group_user.group_id,
-                "group": GroupUserGroupSerializer(group_user).data,
+                "group_user": GroupUserSerializer(group_user).data,
             },
             getattr(user, "web_socket_id", None),
         )
     )
+
+
+@receiver(signals.group_user_updated)
+def group_user_updated(sender, group_user, user, **kwargs):
+    transaction.on_commit(
+        lambda: broadcast_to_group.delay(
+            group_user.group_id,
+            {
+                "type": "group_user_updated",
+                "id": group_user.id,
+                "group_id": group_user.group_id,
+                "group_user": GroupUserSerializer(group_user).data,
+            },
+            getattr(user, "web_socket_id", None),
+        )
+    )
+
+
+@receiver(signals.group_user_deleted)
+def group_user_deleted(sender, group_user_id, group_user, user, **kwargs):
+    def broadcast_to_group_and_removed_user():
+        broadcast_to_group.delay(
+            group_user.group_id,
+            {
+                "type": "group_user_deleted",
+                "id": group_user_id,
+                "group_id": group_user.group_id,
+                "group_user": GroupUserSerializer(group_user).data,
+            },
+            getattr(user, "web_socket_id", None),
+        )
+        broadcast_to_users.delay(
+            [group_user.user_id],
+            {
+                "type": "group_user_deleted",
+                "id": group_user_id,
+                "group_id": group_user.group_id,
+                "group_user": GroupUserSerializer(group_user).data,
+            },
+            getattr(user, "web_socket_id", None),
+        )
+
+    transaction.on_commit(broadcast_to_group_and_removed_user)
 
 
 @receiver(signals.group_restored)
@@ -76,17 +177,6 @@ def group_restored(sender, group_user, user, **kwargs):
                     ).all()
                 ],
             },
-            getattr(user, "web_socket_id", None),
-        )
-    )
-
-
-@receiver(signals.group_user_deleted)
-def group_user_deleted(sender, group_user, user, **kwargs):
-    transaction.on_commit(
-        lambda: broadcast_to_users.delay(
-            [group_user.user_id],
-            {"type": "group_deleted", "group_id": group_user.group_id},
             getattr(user, "web_socket_id", None),
         )
     )